FSA-1107: Moved keyboard navigation code into its own module to make navigation.js less bloated.
